update some sound code

+++ core/src/com/xqs/mypaoku/actor/base/BaseEnemy.java
@@ -164,7 +164,7 @@ public abstract class BaseEnemy extends AnimationActor{
 
         timeCounter+=delta;
 
-        if((float)timeCounter>next){
+        if(timeCounter>next){
             next+=0.1f;
             counter++;
             orderAct(delta,counter);
